Move tests back to csr_kernels
