+++ cuda/test/matrix/CMakeLists.txt
@@ -1,6 +1,5 @@
 ginkgo_create_test(coo_kernels)
 ginkgo_create_test(csr_kernels)
-ginkgo_create_test(csr_kernels2)
 ginkgo_create_test(fbcsr_kernels)
 ginkgo_create_test(dense_kernels)
 ginkgo_create_test(diagonal_kernels)
