--- cuda/test/matrix/csr_kernels2.cpp
@@ -1,195 +1,0 @@
-/*******************************<GINKGO LICENSE>******************************
-Copyright (c) 2017-2021, the Ginkgo authors
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions
-are met:
-
-1. Redistributions of source code must retain the above copyright
-notice, this list of conditions and the following disclaimer.
-
-2. Redistributions in binary form must reproduce the above copyright
-notice, this list of conditions and the following disclaimer in the
-documentation and/or other materials provided with the distribution.
-
-3. Neither the name of the copyright holder nor the names of its
-contributors may be used to endorse or promote products derived from
-this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
-IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
-TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-******************************<GINKGO LICENSE>*******************************/
-
-#include <algorithm>
-#include <numeric>
-#include <random>
-#include <vector>
-
-
-#include <gtest/gtest.h>
-
-
-#include <ginkgo/core/base/exception.hpp>
-#include <ginkgo/core/base/executor.hpp>
-#include <ginkgo/core/matrix/coo.hpp>
-#include <ginkgo/core/matrix/csr.hpp>
-#include <ginkgo/core/matrix/dense.hpp>
-#include <ginkgo/core/matrix/diagonal.hpp>
-#include <ginkgo/core/matrix/identity.hpp>
-#include <ginkgo/core/matrix/sparsity_csr.hpp>
-
-
-#include "core/components/prefix_sum.hpp"
-#include "core/matrix/csr_kernels.hpp"
-#include "core/test/utils.hpp"
-#include "core/test/utils/unsort_matrix.hpp"
-
-
-namespace {
-
-
-class Csr : public ::testing::Test {
-protected:
-    using Arr = gko::Array<int>;
-    using Mtx = gko::matrix::Csr<>;
-    using Vec = gko::matrix::Dense<>;
-
-    Csr()
-#ifdef GINKGO_FAST_TESTS
-        : mtx_size(131, 155),
-#else
-        : mtx_size(532, 231),
-#endif
-          rand_engine(42)
-    {}
-
-    void SetUp()
-    {
-        ref = gko::ReferenceExecutor::create();
-        cuda = gko::CudaExecutor::create(0, ref);
-    }
-
-    void TearDown()
-    {
-        if (cuda != nullptr) {
-            ASSERT_NO_THROW(cuda->synchronize());
-        }
-    }
-
-    template <typename MtxType>
-    std::unique_ptr<MtxType> gen_mtx(int num_rows, int num_cols,
-                                     int min_nnz_row, int max_nnz_row)
-    {
-        return gko::test::generate_random_matrix<MtxType>(
-            num_rows, num_cols,
-            std::uniform_int_distribution<>(min_nnz_row, max_nnz_row),
-            std::normal_distribution<>(-1.0, 1.0), rand_engine, ref);
-    }
-
-    template <typename MtxType>
-    std::unique_ptr<MtxType> gen_mtx(int num_rows, int num_cols,
-                                     int min_nnz_row)
-    {
-        return gen_mtx<MtxType>(num_rows, num_cols, min_nnz_row, num_cols);
-    }
-
-    void set_up_apply_data()
-    {
-        mtx = Mtx::create(ref);
-        mtx->copy_from(gen_mtx<Mtx>(mtx_size[0], mtx_size[1], 5));
-        dmtx = Mtx::create(cuda);
-        dmtx->copy_from(mtx.get());
-    }
-
-    std::shared_ptr<gko::ReferenceExecutor> ref;
-    std::shared_ptr<const gko::CudaExecutor> cuda;
-
-    const gko::dim<2> mtx_size;
-    std::ranlux48 rand_engine;
-    std::unique_ptr<Mtx> mtx;
-    std::unique_ptr<Mtx> dmtx;
-};
-
-
-TEST_F(Csr, CalculateNnzPerRowIsEquivalentToRef)
-{
-    using Mtx = gko::matrix::Csr<>;
-    set_up_apply_data();
-
-    gko::span rspan{7, 51};
-    gko::span cspan{22, 88};
-    auto size = this->mtx->get_size();
-    auto row_nnz = gko::Array<int>(this->ref, rspan.length() + 1);
-    row_nnz.fill(gko::zero<int>());
-    auto drow_nnz = gko::Array<int>(this->cuda, row_nnz);
-
-    gko::kernels::reference::csr::calculate_nonzeros_per_row_in_span(
-        this->ref, this->mtx.get(), rspan, cspan, &row_nnz);
-    gko::kernels::cuda::csr::calculate_nonzeros_per_row_in_span(
-        this->cuda, this->dmtx.get(), rspan, cspan, &drow_nnz);
-
-    GKO_ASSERT_ARRAY_EQ(row_nnz, drow_nnz);
-}
-
-
-TEST_F(Csr, ComputeSubmatrixIsEquivalentToRef)
-{
-    using Mtx = gko::matrix::Csr<>;
-    set_up_apply_data();
-
-    gko::span rspan{7, 51};
-    gko::span cspan{22, 88};
-    auto size = this->mtx->get_size();
-    auto row_nnz = gko::Array<int>(this->ref, rspan.length() + 1);
-    row_nnz.fill(gko::zero<int>());
-    gko::kernels::reference::csr::calculate_nonzeros_per_row_in_span(
-        this->ref, this->mtx.get(), rspan, cspan, &row_nnz);
-    gko::kernels::reference::components::prefix_sum(
-        this->ref, row_nnz.get_data(), row_nnz.get_num_elems());
-    auto drow_nnz = gko::Array<int>(this->cuda, row_nnz);
-    auto smat1 =
-        Mtx::create(this->ref, gko::dim<2>(rspan.length(), cspan.length()),
-                    std::move(row_nnz));
-    auto sdmat1 =
-        Mtx::create(this->cuda, gko::dim<2>(rspan.length(), cspan.length()),
-                    std::move(drow_nnz));
-
-
-    gko::kernels::reference::csr::compute_submatrix(this->ref, this->mtx.get(),
-                                                    rspan, cspan, smat1.get());
-    gko::kernels::cuda::csr::compute_submatrix(this->cuda, this->dmtx.get(),
-                                               rspan, cspan, sdmat1.get());
-    auto hmat = Mtx::create(this->ref);
-    hmat->copy_from(sdmat1.get());
-
-    GKO_ASSERT_MTX_NEAR(hmat, smat1, 0.0);
-}
-
-
-TEST_F(Csr, CreateSubMatrixIsEquivalentToRef)
-{
-    using Mtx = gko::matrix::Csr<>;
-    set_up_apply_data();
-
-    gko::span rspan{47, 81};
-    gko::span cspan{2, 31};
-    auto smat1 = this->mtx->create_submatrix(rspan, cspan);
-    auto sdmat1 = this->dmtx->create_submatrix(rspan, cspan);
-    auto hmat = Mtx::create(this->ref);
-    hmat->copy_from(sdmat1.get());
-
-    GKO_ASSERT_MTX_NEAR(hmat, smat1, 0.0);
-}
-
-
-}  // namespace
